Move fetchAllMiddleware to its own module
